requirements.txt
Flask==1.1.1
numpy==1.21.2
gunicorn==20.1.0
opencv-python-headless==4.5.5.62
Pillow==9.1.1
